+++ tests/testthat/test-posterior.R
@@ -77,4 +77,22 @@ if (.runThisTest && require("posterior") && require("brms")) {
       )
     )
   })
+
+
+  test_that("mp-posterior-draws_rvar", {
+    # Create random vectors by adding an additional dimension:
+    n <- 4 # length of output vector
+    set.seed(123)
+    x <- rvar(array(rnorm(4000 * n, mean = rep(1:n, each = 4000), sd = 1), dim = c(4000, n)))
+    mp <- describe_posterior(x)
+    expect_equal(mp$Median, c(0.99503, 1.99242, 2.9899, 3.99362), tolerance = 1e-2, ignore_attr = TRUE)
+    expect_equal(mp$Parameter, c("x[1]", "x[2]", "x[3]", "x[4]"))
+    expect_equal(
+      colnames(mp),
+      c(
+        "Parameter", "Median", "CI", "CI_low", "CI_high", "pd", "ROPE_CI",
+        "ROPE_low", "ROPE_high", "ROPE_Percentage"
+      )
+    )
+  })
 }